refactor: introduce BaseLocationIndicator
