+++ org.envirocar.map/src/main/java/org/envirocar/map/location/annotation/LocationBearingMarker.kt
@@ -1,4 +1,4 @@
-package org.envirocar.map.location
+package org.envirocar.map.location.annotation
 
 import android.content.Context
 import android.graphics.Bitmap
@@ -9,12 +9,13 @@ import org.envirocar.map.model.Marker
 import org.envirocar.map.model.Point
 
 /**
- *  [LocationPointMarker]
+ *  [LocationBearingMarker]
  *  -------------------------
- *  The [Marker] used to display the current location.
+ *  The [Marker] used to display the current location's bearing.
  */
-internal class LocationPointMarker(
+internal class LocationBearingMarker(
     point: Point,
+    bearing: Float,
     context: Context
 ) : Marker(
     ID,
@@ -22,11 +23,11 @@ internal class LocationPointMarker(
     TITLE,
     DRAWABLE,
     bitmap ?: synchronized(lock) {
-        bitmap ?: AppCompatResources.getDrawable(context, R.drawable.location_point)!!.toBitmap()
-            .also { bitmap = it }
+         bitmap ?: AppCompatResources.getDrawable(context, R.drawable.location_bearing)!!.toBitmap()
+             .also { bitmap = it }
     },
     SCALE,
-    ROTATION
+    bearing
 ) {
 
     companion object {
@@ -35,10 +36,9 @@ internal class LocationPointMarker(
         @Volatile
         private var bitmap: Bitmap? = null
 
-        private const val ID = -0xCL
+        private const val ID = -0xBL
         private val TITLE = null
         private val DRAWABLE = null
-        private const val SCALE = 0.18F
-        private const val ROTATION = 0.0F
+        private const val SCALE = 0.2F
     }
 }
